Move wr js module files to own subdir; bump version of css
