+++ jobqueue/static/js/wr/modal-handlers.js
