+++ jobqueue/static/js/wr/repgroup-handler.js
