+++ jobqueue/static/js/wr/ui-helpers.js
