+++ jobqueue/static/js/wr/utility.js
