+++ jobqueue/static/status.html
@@ -21,9 +21,9 @@
     <script src="/js/knockstrap-1.3.2.min.js"></script>
 
     <!-- Main entry point for the status page (which imports all other modules) -->
-    <script src="/js/main.js" type="module"></script>
+    <script src="/js/wr/main.js" type="module"></script>
 
-    <link rel="stylesheet" href="/css/wr-0.0.1.css">
+    <link rel="stylesheet" href="/css/wr-0.36.0.css">
 </head>
 
 <body>
